rename hkdf to hmac
