+++ src/hmac.rs
