+++ src/lib.rs
@@ -48,8 +48,8 @@ pub mod bulletproofs_tests;
 pub mod aes_tests;
 
 #[cfg(test)]
-#[path = "tests/hkdf_tests.rs"]
-pub mod hkdf_tests;
+#[path = "tests/hmac_tests.rs"]
+pub mod hmac_tests;
 
 // Signing traits
 pub mod traits;
@@ -63,7 +63,7 @@ pub mod bulletproofs;
 pub mod aes;
 
 // Other tooling
-pub mod hkdf;
+pub mod hmac;
 pub mod pubkey_bytes;
 pub mod serde_helpers;
